Added the start of moving fedorahosted's Hosting Requests to an automated process. (See infra ticket 2131). The script doesn't actually execute any commands or such yet, but does get information about tickets, parsed from a template that users are expected to fill out, when they request a project.. Upon mmcgrath returning from PHX and being able to give me orientation, I will play with it on hosted2, and actually make it do stuff ;).
Another thing it does not handle yet is a user incorrectly filling out the template, or deleting questions or such. This needs handled before it gets moved to hosted1 obviously, but I'd like it to actually /work/ first (assuming a correctly filled-in template), and thus be able to test it on hosted2. At this point, I'm just committing it to show that I have started on it, and so there's a copy of it out there, should anything happen. Don't expect it to ...work yet ;).